# cds: cache well known clusters