--- a/pilot/pkg/networking/util/util_test.go
+++ b/pilot/pkg/networking/util/util_test.go
@@ -513,6 +513,42 @@ func TestGetByAddress(t *testing.T) {
 	}
 }
 
+func TestMessageToAnyCache(t *testing.T) {
+	// Call MessageToAny to populate the cache.
+	cluster := MessageToAny(buildFakeCluster())
+
+	if !MessageCache.Contains(buildFakeCluster().String()) {
+		t.Fatal("Expected to be in cache, but not available")
+	}
+
+	if MessageToAny(buildFakeCluster()) != cluster {
+		t.Fatal("Cache is not returning the same any structure")
+	}
+
+}
+
+func BenchmarkMessageToAny(b *testing.B) {
+	// inHCM := &http_conn.HttpConnectionManager{
+	// 	CodecType:  http_conn.HttpConnectionManager_HTTP1,
+	// 	StatPrefix: "123",
+	// 	HttpFilters: []*http_conn.HttpFilter{
+	// 		{
+	// 			Name: "filter1",
+	// 			ConfigType: &http_conn.HttpFilter_TypedConfig{
+	// 				TypedConfig: &any.Any{},
+	// 			},
+	// 		},
+	// 	},
+	// 	ServerName:        "scooby",
+	// 	XffNumTrustedHops: 2,
+	// }
+	c := buildFakeCluster()
+	for n := 0; n < b.N; n++ {
+	//	MessageToAny(inHCM)
+		MessageToAny(c)
+	}
+}
+
 func TestMergeAnyWithStruct(t *testing.T) {
 	inHCM := &http_conn.HttpConnectionManager{
 		CodecType:  http_conn.HttpConnectionManager_HTTP1,