Currently, we build standard well known clusters like Blackhole cluster and Passthrough clusters every time for every proxy. We can build them once, and reuse them whenever CDS is being built. This PR implements cache of these well known clusters.

[ ] Configuration Infrastructure
[ ] Docs
[ ] Installation
[X ] Networking
[ ] Performance and Scalability
[ ] Policies and Telemetry
[ ] Security
[ ] Test and Release
[ ] User Experience
[ ] Developer Infrastructure