@istio-testing 2019-12-29
@ramaraochavali: The following tests **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
gencheck_istio | c2515081fdf57362550056815781065345a33f95 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/19842/gencheck_istio/4011) | `/test gencheck_istio`
lint_istio | c2515081fdf57362550056815781065345a33f95 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/19842/lint_istio/5486) | `/test lint_istio`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->