@ramaraochavali 2019-12-29
Benchmark results of `BuildClusters` 

Before    	   59734	     17394 ns/op	   14797 B/op	     229 allocs/op
After  	   74468	     15992 ns/op	   11935 B/op	     202 allocs/op

Benchmark results of 'Anywithcache`

Before    	 1370701	       874 ns/op	     272 B/op	       3 allocs/op
After    	27717297      43.5 ns/op	       0 B/op	       0 allocs/op