@ramaraochavali 2019-12-29
Yeah. Only connect timeout is non static. I am thinking about how to reset them when that config is changed - one way I am thinking is to hold the connect timeout in a var and if it
changes reset these clusters - any other better way?