@howardjohn 2019-12-29
And if we assume it's just mesh config that impacts this, we could just
clear the cache on mesh config updates potentially

On Sun, Dec 29, 2019, 7:31 AM John Howard <howardjohn@google.com> wrote:

> What performance improvement in a real cluster are we seeing with this?
> adding all this caching is a slippery slope of complexity and subtle bugs
> so if we do it as should make sure it's worth it.
>
> On Sun, Dec 29, 2019, 7:20 AM Rama Chavali <notifications@github.com>
> wrote:
>
>> Yeah. Only connect timeout is non static. I am thinking about how to
>> reset them when that config is changed - one way I am thinking is to hold
>> the connect timeout in a var and if it
>> changes reset these clusters - any other better way?
>>
>> —
>> You are receiving this because your review was requested.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/istio/istio/pull/19842?email_source=notifications&email_token=AAEYGXLIDKYUXPK62CGFQZ3Q3C53TA5CNFSM4KAYD2B2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEHZBXXA#issuecomment-569514972>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AAEYGXJ4IGE7IF6SUKU22O3Q3C53TANCNFSM4KAYD2BQ>
>> .
>>
>