@ramaraochavali 2019-12-30
> What performance improvement in a real cluster are we seeing with this?

I was not able on a big enough cluster to validate how signigicant this performance improvement is. Was just trying with basic benchmarks.

> adding all this caching is a slippery slope of complexity and subtle bugs
so if we do it as should make sure it's worth it.

Completely agree with this - given the dependency on mesh config (and probably some other things in future), It can lead to subtle bugs.