@ramaraochavali 2019-12-30
I am closing this for now, will open again once I test and confirm the perf improvement on real big enough cluster.